+++ src/script_interpreter_func_fe3d_quad2d_get.cpp
@@ -261,7 +261,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dGetter(const string & functionNa
 		{
 			if(_validateFe3dQuad2d(args[0]->getString(), false))
 			{
-				const auto result = static_cast<int>(_fe3d->quad2d_getTextureRepeat(args[0]->getString()));
+				const auto result = _fe3d->quad2d_getTextureRepeat(args[0]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -509,7 +509,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dGetter(const string & functionNa
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->quad2d_getAnimationRowIndex(args[0]->getString(), args[1]->getString()));
+				const auto result = _fe3d->quad2d_getAnimationRowIndex(args[0]->getString(), args[1]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -529,7 +529,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dGetter(const string & functionNa
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->quad2d_getAnimationColumnIndex(args[0]->getString(), args[1]->getString()));
+				const auto result = _fe3d->quad2d_getAnimationColumnIndex(args[0]->getString(), args[1]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -569,7 +569,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dGetter(const string & functionNa
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->quad2d_getAnimationIntervalMultiplier(args[0]->getString(), args[1]->getString()));
+				const auto result = _fe3d->quad2d_getAnimationIntervalMultiplier(args[0]->getString(), args[1]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -589,7 +589,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dGetter(const string & functionNa
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->quad2d_getAnimationIntervalDivider(args[0]->getString(), args[1]->getString()));
+				const auto result = _fe3d->quad2d_getAnimationIntervalDivider(args[0]->getString(), args[1]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
