+++ src/script_interpreter_func_fe3d_quad3d_get.cpp
@@ -362,7 +362,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dGetter(const string & functionNa
 		{
 			if(_validateFe3dQuad3d(args[0]->getString(), false))
 			{
-				const auto result = static_cast<int>(_fe3d->quad3d_getTextureRepeat(args[0]->getString()));
+				const auto result = _fe3d->quad3d_getTextureRepeat(args[0]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -725,7 +725,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dGetter(const string & functionNa
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->quad3d_getAnimationRowIndex(args[0]->getString(), args[1]->getString()));
+				const auto result = _fe3d->quad3d_getAnimationRowIndex(args[0]->getString(), args[1]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -745,7 +745,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dGetter(const string & functionNa
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->quad3d_getAnimationColumnIndex(args[0]->getString(), args[1]->getString()));
+				const auto result = _fe3d->quad3d_getAnimationColumnIndex(args[0]->getString(), args[1]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -785,7 +785,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dGetter(const string & functionNa
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->quad3d_getAnimationIntervalMultiplier(args[0]->getString(), args[1]->getString()));
+				const auto result = _fe3d->quad3d_getAnimationIntervalMultiplier(args[0]->getString(), args[1]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -805,7 +805,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dGetter(const string & functionNa
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->quad3d_getAnimationIntervalDivider(args[0]->getString(), args[1]->getString()));
+				const auto result = _fe3d->quad3d_getAnimationIntervalDivider(args[0]->getString(), args[1]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
