+++ src/script_interpreter_func_fe3d_sound2d_get.cpp
@@ -178,7 +178,7 @@ const bool ScriptInterpreter::_executeFe3dSound2dGetter(const string & functionN
 					return true;
 				}
 
-				const auto result = static_cast<int>(_fe3d->sound2d_getTime(args[0]->getString(), args[1]->getInteger()));
+				const auto result = _fe3d->sound2d_getTime(args[0]->getString(), args[1]->getInteger());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
@@ -192,7 +192,7 @@ const bool ScriptInterpreter::_executeFe3dSound2dGetter(const string & functionN
 		{
 			if(_validateFe3dSound2d(args[0]->getString(), false))
 			{
-				const auto result = static_cast<int>(_fe3d->sound2d_getStartedCount(args[0]->getString()));
+				const auto result = _fe3d->sound2d_getStartedCount(args[0]->getString());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
